put in a price check/adder to the html list
Added a checker if the price isn't the one you set then it does add it to the list, next to add is a while loop true to have it going continuously as its open. need to make a setting txt for each email which will include high price low price the searched item with which item has already been sent to your email. then hopefully make another version which has tkinter and isnt functional. using classes